+++ source/common/http/BUILD
@@ -160,9 +160,9 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
-    name = "alternate_protocols",
-    srcs = ["alternate_protocols.cc"],
-    hdrs = ["alternate_protocols.h"],
+    name = "alternate_protocols_cache",
+    srcs = ["alternate_protocols_cache.cc"],
+    hdrs = ["alternate_protocols_cache.h"],
     deps = [
         "//include/envoy/common:time_interface",
     ],
@@ -173,7 +173,7 @@ envoy_cc_library(
     srcs = ["conn_pool_grid.cc"],
     hdrs = ["conn_pool_grid.h"],
     deps = [
-        ":alternate_protocols",
+        ":alternate_protocols_cache",
         ":http3_status_tracker",
         ":mixed_conn_pool",
         "//source/common/http/http3:conn_pool_lib",
